% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfTournWinners.R
\name{getConfTournWinners}
\alias{getConfTournWinners}
\title{Get Conference Tournament Winners}
\usage{
getConfTournWinners(stage = 1)
}
\arguments{
\item{stage}{Either 1 or 2, denotes the stage of the competition to use data from. Default is 1.}
}
\value{
A dataframe containing the conference tournament winners for each year in the dataset
}
\description{
Get Conference Tournament Winners
}
\examples{
#get conference tournament winners from stage 2 data
\dontrun{
confTournWinners <- getConfTournWinners(stage = 2)
}

}
