% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCAAMData.R
\name{getNCAAMData}
\alias{getNCAAMData}
\title{Download and/or Read in Kaggle's NCAA Data}
\usage{
getNCAAMData()
}
\value{
Vector of dataset names read in, as well as dataframes in the global environment for selected data. All datasets will also be downloaded to the current directory if not already there.
}
\description{
Function to download Kaggle's 2021 March Mania competition data (if not already downloaded) and/or read in selected datasets
}
