% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRanks}
\alias{getRanks}
\title{Get Summarized Rankings of Teams}
\usage{
getRanks(sysName = "all", stage = 1, ...)
}
\arguments{
\item{sysName}{The ranking system name(s) to filter the data by. Defaults to all, so all systems are used.}

\item{stage}{Either 1 or 2, denotes the stage of the competition to use data from. Default is 1.}

\item{...}{Unquoted function names to use to summarize the ranking statistics, such as mean, median, or standard deviation. Will use mean if nothing is specified.}
}
\value{
A dataframe containing summarized rankings for each team in each season.
}
\description{
Get Summarized Rankings of Teams
}
\examples{
#get mean rankings based on all systems in stage 1
\dontrun{
rankings <- getRanks()
}

#get mean and standard deviation of rankings based on the Massey (MAS) rankings and AP rankings
\dontrun{
rankings2 <- getRanks(sysName = c("MAS", "AP"), stage = 1, mean, sd)
}

}
