% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecords.R
\name{getRecords}
\alias{getRecords}
\title{Get Different Records for Teams}
\usage{
getRecords(stage = 1)
}
\arguments{
\item{stage}{Either 1 or 2, denotes the stage of the competition to use data from. Default is 1.}
}
\value{
A dataframe containing wins, losses, and win percentages for various types of games (home, away, neutral) for each team in each season
}
\description{
Get Different Records for Teams
}
\examples{
#get stage 2 records data
\dontrun{
records <- getRecords(stage = 2)
}

}
