% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeeds.R
\name{getSeeds}
\alias{getSeeds}
\title{Get Seeds for Tournament Teams}
\usage{
getSeeds(stage = 1)
}
\arguments{
\item{stage}{Either 1 or 2, denotes the stage of the competition to use data from. Default is 1.}
}
\value{
A dataframe containing the seeds as integers for each tournament team in each season
}
\description{
Get Seeds for Tournament Teams
}
\examples{
#get tournament seeds for stage 2 data
\dontrun{
seeds <- getSeeds(stage = 2)
}

}
