% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottaStats.R
\name{lottaStats}
\alias{lottaStats}
\title{Put Other Statistic Functions Together...and Add a Little More}
\usage{
lottaStats(stage = 1, rankSysName = "all")
}
\arguments{
\item{stage}{Either 1 or 2, denotes the stage of the competition to use data from. Default is 1.}

\item{rankSysName}{The ranking system name(s) to filter the data by. Defaults to all, so all systems are used.}
}
\value{
A dataframe containing average per-game statistics, average rankings, conference, team name, seed (NA if they didn't make the tournament), and if the team won the conference tournament for each team in each season.
}
\description{
Put Other Statistic Functions Together...and Add a Little More
}
\examples{
#get all the variables using stage 2 data and all ranking systems
\dontrun{
bigStats <- lottaStats(stage = 2, rankSysName = "all")
}

#use stage 1 data and only the Massey (MAS) rankings
\dontrun{
bigStats2 <- lottaStats(stage = 1, rankSysName = "MAS")
}

}
