% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaaBracket.R
\name{ncaaBracket}
\alias{ncaaBracket}
\title{Fill in Your NCAA Bracket}
\usage{
ncaaBracket(season = 2021, filled = FALSE, fillData = NULL)
}
\arguments{
\item{season}{The season for the tournament. Defaults to 2021.}

\item{filled}{A logical value denoting whether to fill in winners. If filled = FALSE, the bracket will only show the first round with every team included. Defaults to FALSE.}

\item{fillData}{A vector containing your predicted winners for each game. The order should be: first round winners on left side of bracket from top to bottom, first round winners on right side of bracket from top to bottom, and so on for each round, with the champion being last. See an example below for an actual idea. If you choose a season with play-in games and want to predict the play-in games, please list those winners first. It would be beneficial to look at the unfilled bracket first so you know the correct order.}
}
\value{

}
\description{
Fill in Your NCAA Bracket
}
\examples{
#get 2021 bracket, not filled in
\dontrun{
ncaaBracket()
}

#get the 2014 bracket filled in with the actual winners
\dontrun{
ncaaBracket(season = 2014, filled = TRUE)
}

#fill in the 2015 bracket with your own predictions
#first four are my predicted play-in winners, but you can exclude them if wanted
\dontrun{
predictions <- c("Hampton", "Dayton", "North Florida", "Ole Miss", "Kentucky", "Purdue", "West Virginia", "Maryland",
                 "Texas", "Notre Dame", "Indiana", "Kansas", "Wisconsin", "Oregon", "Arkansas", "North Carolina",
                 "Xavier", "Baylor", "Ohio State", "Arizona", "Villanova", "NC State", "Northern Iowa", "Louisville",
                 "Dayton", "Oklahoma", "Michigan State", "Virginia", "Duke", "St. John's", "Utah", "Georgetown",
                 "UCLA", "Iowa State", "Iowa", "Gonzaga", "Kentucky", "West Virginia", "Notre Dame", "Kansas",
                 "Wisconsin", "Arkansas", "Baylor", "Arizona", "Villanova", "Louisville", "Oklahoma", "Michigan State",
                 "Duke", "Utah", "Iowa State", "Gonzaga", "Kentucky", "Notre Dame", "Wisconsin", "Arizona", "Villanova",
                 "Michigan State", "Duke", "Iowa State", "Kentucky", "Wisconsin", "Michigan State", "Duke", "Kentucky",
                 "Duke", "Duke")

ncaaBracket(season = 2015, filled = TRUE, fillData = predictions)
}


}
