% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perGameStats.R
\name{perGameStats}
\alias{perGameStats}
\title{Get Per-Game Statistics}
\usage{
perGameStats(stage = 1, ...)
}
\arguments{
\item{stage}{Either 1 or 2, denotes the stage of the competition to use data from. Default is 1.}

\item{...}{Unquoted function names to use to summarize the per-game statistics, such as mean, median, or standard deviation. Will use mean if nothing is specified.}
}
\value{
A dataframe containing summarized statistics for each team in each season.
}
\description{
Get Per-Game Statistics
}
\examples{
#will use stage 1 data and calculate the mean for each of the statistics
\dontrun{
avgStatsPerGame <- perGameStats()
}

#will use stage 2 data and calculate the mean, median, and standard deviation of each statistic
\dontrun{
perGameStats(stage = 2, mean, median, sd)
}

}
